move app store in app folder
