+++ src/app/pages/admin/admin.module.ts
@@ -3,8 +3,8 @@ import { CommonModule } from '@angular/common';
 import { SharedModule } from '@shared/shared.module';
 import { AdminComponent } from './admin.component';
 import { adminRoutes } from './admin.routes';
-import { MaterialComponentsModule } from 'src/app/shared/material.module';
-import { HeaderComponent } from 'src/app/shared/header/header.component';
+import { MaterialComponentsModule } from '@shared/material.module';
+import { HeaderComponent } from '@shared/header/header.component';
 import { RouterModule } from '@angular/router';
 
 @NgModule({
