+++ src/app/pages/admin/beneficiaries/beneficiaries.facade.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Store, select } from '@ngrx/store';
 
-import { AppState } from '@store/app.state';
+import { AppState } from '@app/app.state';
 
 import {
   saveBeneficiaryAction,
